log thread info

+++ src/utils/constant.h
@@ -9,8 +9,8 @@
 
 #define MPI_CREAL MPI_FLOAT
 
-//typedef float real;
-typedef double real;
+typedef float real;
+//typedef double real;
 
 typedef int ID;
